backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LLM providers
openai==1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
httpx==0.26.0

# Tools
beautifulsoup4==4.12.3
duckduckgo-search==4.4.3
aiofiles==23.2.1

# Advanced Tools
playwright==1.41.0
PyPDF2==3.0.1
pillow==10.2.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
requests==2.31.0

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Scheduling
apscheduler==3.10.4

# Email
aiosmtplib==3.0.1

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
rich==13.7.0
networkx==3.2.1
pyyaml>=6.0

# Security & Reliability
simpleeval>=0.9.13
tenacity>=8.2.0
email-validator>=2.1.0

# Sandbox (optional)
docker>=7.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
